Datasets

+++ genetica-libraries/src/test/java/umcg/genetica/math/matrix2/DoubleMatrixDatasetTest.java
@@ -46,13 +46,15 @@ public class DoubleMatrixDatasetTest {
 		Runtime.getRuntime().addShutdownHook(new Thread() {
 			@Override
 			public void run() {
-				System.out.println("Removing tmp dir and files");
-				for (File file : tmpOutputFolder.listFiles()) {
-					System.out.println(" - Deleting: " + file.getAbsolutePath());
-					file.delete();
+				if (tmpOutputFolder.isDirectory()) {
+					System.out.println("Removing tmp dir and files");
+					for (File file : tmpOutputFolder.listFiles()) {
+						System.out.println(" - Deleting: " + file.getAbsolutePath());
+						file.delete();
+					}
+					System.out.println(" - Deleting: " + tmpOutputFolder.getAbsolutePath());
+					tmpOutputFolder.delete();
 				}
-				System.out.println(" - Deleting: " + tmpOutputFolder.getAbsolutePath());
-				tmpOutputFolder.delete();
 			}
 		});
 
@@ -155,26 +157,26 @@ public class DoubleMatrixDatasetTest {
 		assertEquals(dataset3.getRowObjects().get(0), "row3");
 		assertEquals(dataset3.getRowObjects().get(1), "row2");
 		assertEquals(dataset3.getColObjects().get(2), "col3");
-		
+
 		assertEquals(dataset3.getElementQuick(0, 2), 6.66d);
 		assertEquals(dataset3.getElementQuick(0, 3), -12.2d);
 		assertEquals(dataset3.getElementQuick(1, 3), 5.55d);
 		assertEquals(dataset3.getElementQuick(1, 2), 0d);
-		
+
 		DoubleMatrixDataset<String, String> dataset4 = dataset2.viewRowSelection(new String[]{"row3", "row2"});
-		
+
 		assertEquals(dataset4.rows(), 2);
 		assertEquals(dataset4.columns(), 5);
 
 		assertEquals(dataset4.getRowObjects().get(0), "row3");
 		assertEquals(dataset4.getRowObjects().get(1), "row2");
 		assertEquals(dataset4.getColObjects().get(2), "col3");
-		
+
 		assertEquals(dataset4.getElementQuick(0, 2), 6.66d);
 		assertEquals(dataset4.getElementQuick(0, 3), -12.2d);
 		assertEquals(dataset4.getElementQuick(1, 3), 5.55d);
 		assertEquals(dataset4.getElementQuick(1, 2), 0d);
-		
+
 		DoubleMatrixDatasetFastSubsetLoader subsetLoader = new DoubleMatrixDatasetFastSubsetLoader(tmpOutputFolder.getAbsolutePath() + ".testBin");
 		DoubleMatrixDataset<String, String> dataset5 = subsetLoader.loadSubsetOfRowsBinaryDoubleData(new String[]{"row3", "row2"});
 
@@ -184,15 +186,14 @@ public class DoubleMatrixDatasetTest {
 		assertEquals(dataset5.getRowObjects().get(0), "row3");
 		assertEquals(dataset5.getRowObjects().get(1), "row2");
 		assertEquals(dataset5.getColObjects().get(2), "col3");
-		
+
 		assertEquals(dataset5.getElementQuick(0, 2), 6.66d);
 		assertEquals(dataset5.getElementQuick(0, 3), -12.2d);
 		assertEquals(dataset5.getElementQuick(1, 3), 5.55d);
 		assertEquals(dataset5.getElementQuick(1, 2), 0d);
-		
+
 	}
-	
-	
+
 	@Test
 	public void testSaveLoadTextMatrix() throws IOException, Exception {
 
@@ -239,8 +240,7 @@ public class DoubleMatrixDatasetTest {
 		HashSet<String> rowsToLoad = new HashSet<>();
 		rowsToLoad.add("row3");
 		rowsToLoad.add("row2");
-		
-				
+
 		DoubleMatrixDataset<String, String> dataset3 = DoubleMatrixDataset.loadSubsetOfTextDoubleData(tmpOutputFolder.getAbsolutePath() + ".testText.txt", '\t', rowsToLoad, null);
 
 		assertEquals(dataset3.rows(), 2);
@@ -249,32 +249,30 @@ public class DoubleMatrixDatasetTest {
 		assertEquals(dataset3.getRowObjects().get(0), "row2");
 		assertEquals(dataset3.getRowObjects().get(1), "row3");
 		assertEquals(dataset3.getColObjects().get(2), "col3");
-		
+
 		assertEquals(dataset3.getElementQuick(1, 2), 6.66d);
 		assertEquals(dataset3.getElementQuick(1, 3), -12.2d);
 		assertEquals(dataset3.getElementQuick(0, 3), 5.55d);
 		assertEquals(dataset3.getElementQuick(0, 2), 0d);
-		
+
 		DoubleMatrixDataset<String, String> dataset4 = dataset3.viewRowSelection(new String[]{"row3", "row2"});
-		
+
 		assertEquals(dataset4.rows(), 2);
 		assertEquals(dataset4.columns(), 5);
 
 		assertEquals(dataset4.getRowObjects().get(0), "row3");
 		assertEquals(dataset4.getRowObjects().get(1), "row2");
 		assertEquals(dataset4.getColObjects().get(2), "col3");
-		
+
 		assertEquals(dataset4.getElementQuick(0, 2), 6.66d);
 		assertEquals(dataset4.getElementQuick(0, 3), -12.2d);
 		assertEquals(dataset4.getElementQuick(1, 3), 5.55d);
 		assertEquals(dataset4.getElementQuick(1, 2), 0d);
-		
-		
+
 		HashSet<String> colsToLoad = new HashSet<>();
 		colsToLoad.add("col2");
 		colsToLoad.add("col4");
-		
-		
+
 		DoubleMatrixDataset<String, String> dataset5 = DoubleMatrixDataset.loadSubsetOfTextDoubleData(tmpOutputFolder.getAbsolutePath() + ".testText.txt", '\t', rowsToLoad, colsToLoad);
 
 		assertEquals(dataset5.rows(), 2);
@@ -284,14 +282,14 @@ public class DoubleMatrixDatasetTest {
 		assertEquals(dataset5.getRowObjects().get(1), "row3");
 		assertEquals(dataset5.getColObjects().get(0), "col2");
 		assertEquals(dataset5.getColObjects().get(1), "col4");
-		
+
 		assertEquals(dataset5.getElementQuick(0, 0), 0d);
 		assertEquals(dataset5.getElementQuick(0, 1), 5.55d);
 		assertEquals(dataset5.getElementQuick(1, 0), 0d);
 		assertEquals(dataset5.getElementQuick(1, 1), -12.2d);
-		
+
 		DoubleMatrixDataset<String, String> dataset6 = dataset3.viewColSelection(new String[]{"col2", "col4"});
-		
+
 		assertEquals(dataset6.rows(), 2);
 		assertEquals(dataset6.columns(), 2);
 
@@ -299,14 +297,14 @@ public class DoubleMatrixDatasetTest {
 		assertEquals(dataset6.getRowObjects().get(1), "row3");
 		assertEquals(dataset6.getColObjects().get(0), "col2");
 		assertEquals(dataset6.getColObjects().get(1), "col4");
-		
+
 		assertEquals(dataset6.getElementQuick(0, 0), 0d);
 		assertEquals(dataset6.getElementQuick(0, 1), 5.55d);
 		assertEquals(dataset6.getElementQuick(1, 0), 0d);
 		assertEquals(dataset6.getElementQuick(1, 1), -12.2d);
-		
+
 		DoubleMatrixDataset<String, String> dataset7 = dataset3.viewColSelection(new String[]{"col4", "col2"});
-		
+
 		assertEquals(dataset7.rows(), 2);
 		assertEquals(dataset7.columns(), 2);
 
@@ -314,12 +312,11 @@ public class DoubleMatrixDatasetTest {
 		assertEquals(dataset7.getRowObjects().get(1), "row3");
 		assertEquals(dataset7.getColObjects().get(0), "col4");
 		assertEquals(dataset7.getColObjects().get(1), "col2");
-		
+
 		assertEquals(dataset7.getElementQuick(0, 1), 0d);
 		assertEquals(dataset7.getElementQuick(0, 0), 5.55d);
 		assertEquals(dataset7.getElementQuick(1, 1), 0d);
 		assertEquals(dataset7.getElementQuick(1, 0), -12.2d);
-
 
 	}
